Make handshake less strict with ts and payload
According to the RTMP specification, client's ACK message (C2) must include
the timestamp and payload from server's SYN message (C1). After some
observations done with a Wowza publisher against an Adobe Media Server,
we concluded that those requirements are completely ignored by both
sides and, even worse, Wowza publisher sends completelly bogus data and
timestamps in all the messages involved with the handshake breaking any
attempt to follow the specification.

Having this in mind, we decided to remove the hard check in the
ServerNegotiator to allow publishers like Wowza to still work even if
they fuck with the handshake.