Don't require client's BW negotiation to connect
In some cases, the client may not send the bandwidth negotiation
packets to the server. If this is the case, the server can be stuck
waiting for it and never completing the connect request. An example of
such publisher is the Wowza server.

This patch adds a timeout of 0.1 seconds to the wait for the bandwidth
negotiation packets from the client. If it is reached, the server will
use the default bandwidth and resume the connection as usual.

Also, initialize the _nextInterval private attribute of the decoder,
which was only initialized after receiving the bandwidth negotiation.